requirements.txt
streamlit
python-dotenv
pillow
google-generativeai
pandas
xlsxwriter
spacy
groq
fastapi
uvicorn
python-multipart